backend/employee/filters.py
import django_filters

from .models import Employee


class EmployeeFilter(django_filters.FilterSet):
    email = django_filters.CharFilter(field_name="email", lookup_expr="icontains")
    name = django_filters.CharFilter(field_name="name", lookup_expr="icontains")

    class Meta:
        model = Employee
        fields = ["email", "name"]
