backend/pyproject.toml
[tool.poetry]
name = "config"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.2"
djangorestframework = "^3.15"
django-filter = "^25.1"
django-cors-headers = "^4.7.0"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
black = "^25.1.0"
pytest-django = "^4.11.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I", "W", "UP", "B"]
line-length = 120
fix = true
show-fixes = true
exclude = ["migrations", ".venv", "__pycache__"]

[tool.black]
line-length = 120
target-version = ["py310"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
  | migrations
  | __pycache__
)/
'''
